Import version 4.2.0 from SVN

+++ includes/Core/Supports/Utils/Date_Time.php
@@ -35,7 +35,7 @@ class Date_Time {
 		}
 
 		if ( is_string( $date ) ) {
-			$date = strtotime( $date );
+			$date = (int) strtotime( $date );
 		}
 
 		if ( $gmt_offset ) {
